test: move inspector churn test to its own file
inspect.test.ts has pre-existing failures on main that are unrelated to
this change; keep this test in a separate file so it can be judged on
its own.
